+++ examples/03-Lodging.php
@@ -62,21 +62,22 @@ echo sprintf(
     $lodging->getReceptionAddress()->getLocality()
 ), PHP_EOL;
 
-echo sprintf('   • Contact point     : t %s',
-    $lodging->getContactPoint()->getPhoneNumber()
+echo sprintf(
+    '   • Contact point     : t %s',
+    (string)  $lodging->getContactPoint()->getPhoneNumbers()
 ), PHP_EOL;
 echo sprintf(
     '                         m %s',
-    $lodging->getContactPoint()->getEmailAddress()
+    (string) $lodging->getContactPoint()->getEmailAddresses()
 ), PHP_EOL;
 echo sprintf(
     '                         w %s',
-    $lodging->getContactPoint()->getWebsiteAddress()
+    (string) $lodging->getContactPoint()->getWebsiteAddresses()
 ), PHP_EOL;
 
 echo sprintf('   • Type              : %s', $lodging->getRegistration()->getType()), PHP_EOL;
 echo sprintf('   • Status            : %s', $lodging->getRegistration()->getStatus()), PHP_EOL;
-echo sprintf('   • Star rating       : %s', $lodging->getStarRating()), PHP_EOL;
+echo sprintf('   • Star rating       : %s', $lodging->getRating()), PHP_EOL;
 echo sprintf('   • Quality labels    : %s', $lodging->getQualityLabels()), PHP_EOL;
 echo sprintf('   • Images            : %s', $lodging->getImages()), PHP_EOL;