+++ src/Normalizer/FromJson/ContactInfoNormalizer.php
@@ -4,12 +4,13 @@ declare(strict_types=1);
 
 namespace DigipolisGent\Toerismevlaanderen\Lodging\Normalizer\FromJson;
 
-use DigipolisGent\Toerismevlaanderen\Lodging\Value\Address;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\ContactInfo;
-use DigipolisGent\Toerismevlaanderen\Lodging\Value\Coordinates;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\EmailAddress;
+use DigipolisGent\Toerismevlaanderen\Lodging\Value\EmailAddresses;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\PhoneNumber;
+use DigipolisGent\Toerismevlaanderen\Lodging\Value\PhoneNumbers;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\WebsiteAddress;
+use DigipolisGent\Toerismevlaanderen\Lodging\Value\WebsiteAddresses;
 
 /**
  * Normalizer to get the contact info out of json decoded data.
@@ -27,25 +28,90 @@ class ContactInfoNormalizer
      */
     public function normalize(object $lodgingData, string $prefix): ContactInfo
     {
-        $phoneVariable = sprintf('%s_phoneNumber', $prefix);
-        $phoneNumber = !empty($lodgingData->{$phoneVariable}->value)
-            ? PhoneNumber::fromNumber($lodgingData->{$phoneVariable}->value)
-            : PhoneNumber::withoutNumber();
+        $phoneVariable = sprintf('%s_phoneNumbers', $prefix);
+        $phoneNumbers = $this->normalizePhoneNumbers($lodgingData->{$phoneVariable}->value ?? '');
 
-        $emailVariable = sprintf('%s_emailAddress', $prefix);
-        $emailAddress = !empty($lodgingData->{$emailVariable}->value)
-            ? EmailAddress::fromAddress($lodgingData->{$emailVariable}->value)
-            : EmailAddress::withoutAddress();
+        $emailVariable = sprintf('%s_emailAddresses', $prefix);
+        $emailAddresses = $this->normalizeEmailAddresses($lodgingData->{$emailVariable}->value ?? '');
 
-        $websiteVariable = sprintf('%s_websiteAddress', $prefix);
-        $websiteAddress = !empty($lodgingData->{$websiteVariable}->value)
-            ? WebsiteAddress::fromUrl($lodgingData->{$websiteVariable}->value)
-            : WebsiteAddress::withoutUrl();
+        $websiteVariable = sprintf('%s_websiteAddresses', $prefix);
+        $websiteAddresses = $this->normalizeWebsiteAddresses($lodgingData->{$websiteVariable}->value ?? '');
 
         return ContactInfo::fromDetails(
-            $phoneNumber,
-            $emailAddress,
-            $websiteAddress
+            $phoneNumbers,
+            $emailAddresses,
+            $websiteAddresses
+        );
+    }
+
+    /**
+     * Create the phone numbers collection.
+     *
+     * @param string $data
+     *
+     * @return \DigipolisGent\Toerismevlaanderen\Lodging\Value\PhoneNumbers
+     */
+    private function normalizePhoneNumbers(string $data): PhoneNumbers
+    {
+        $numbers = $this->extractValuesFromString($data);
+        $phoneNumbers = [];
+        foreach ($numbers as $number) {
+            $phoneNumbers[] = PhoneNumber::fromNumber($number);
+        }
+
+        return PhoneNumbers::fromPhoneNumbers(...$phoneNumbers);
+    }
+
+    /**
+     * Create the email addresses collection.
+     *
+     * @param string $data
+     *
+     * @return \DigipolisGent\Toerismevlaanderen\Lodging\Value\EmailAddresses
+     */
+    private function normalizeEmailAddresses(string $data): EmailAddresses
+    {
+        $addresses = $this->extractValuesFromString($data);
+        $emailAddresses = [];
+        foreach ($addresses as $address) {
+            $emailAddresses[] = EmailAddress::fromAddress($address);
+        }
+
+        return EmailAddresses::fromEmailAddresses(...$emailAddresses);
+    }
+
+    /**
+     * Create the website addresses collection.
+     *
+     * @param string $data
+     *
+     * @return \DigipolisGent\Toerismevlaanderen\Lodging\Value\WebsiteAddresses
+     */
+    private function normalizeWebsiteAddresses(string $data): WebsiteAddresses
+    {
+        $urls = $this->extractValuesFromString($data);
+        $websiteAddresses = [];
+        foreach ($urls as $url) {
+            $websiteAddresses[] = WebsiteAddress::fromUrl($url);
+        }
+
+        return WebsiteAddresses::fromWebsiteAddresses(...$websiteAddresses);
+    }
+
+    /**
+     * Get the individual values from field with comma-separated values.
+     *
+     * @param string $value
+     *
+     * @return array
+     */
+    private function extractValuesFromString(string $value): array
+    {
+        return array_filter(
+            array_map(
+                'trim',
+                explode(',', $value)
+            )
         );
     }
 }