+++ src/Serializer/LodgingArraySerializer.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace DigipolisGent\Toerismevlaanderen\Lodging\Serializer;
 
-use DigipolisGent\Toerismevlaanderen\Lodging\Value\Lodging;
+use DigipolisGent\Toerismevlaanderen\Lodging\Value\LodgingInterface;
 
 /**
  * Serializes a Lodging value into an array.
@@ -14,11 +14,11 @@ final class LodgingArraySerializer
     /**
      * Serializes a given Lodging value into an array of data.
      *
-     * @param \DigipolisGent\Toerismevlaanderen\Lodging\Value\Lodging $lodging
+     * @param \DigipolisGent\Toerismevlaanderen\Lodging\Value\LodgingInterface $lodging
      *
      * @return array
      */
-    public function serialize(Lodging $lodging): array
+    public function serialize(LodgingInterface $lodging): array
     {
         $addressSerializer = new AddressArraySerializer();
         $contactInfoSerializer = new ContactInfoArraySerializer();
@@ -35,7 +35,7 @@ final class LodgingArraySerializer
             ],
             'receptionAddress' => $addressSerializer->serialize($lodging->getReceptionAddress()),
             'contactPoint' => $contactInfoSerializer->serialize($lodging->getContactPoint()),
-            'starRating' => (string) $lodging->getStarRating(),
+            'rating' => (string) $lodging->getRating(),
             'qualityLabels' => $lodging->getQualityLabels()->getLabels(),
             'images' => $imagesSerializer->serialize($lodging->getImages()),
         ];