+++ tests/Handler/LodgingHandlerTest.php
@@ -12,15 +12,18 @@ use DigipolisGent\Toerismevlaanderen\Lodging\Value\Address;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\ContactInfo;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\Coordinates;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\EmailAddress;
+use DigipolisGent\Toerismevlaanderen\Lodging\Value\EmailAddresses;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\Image;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\Images;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\Lodging;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\LodgingId;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\PhoneNumber;
+use DigipolisGent\Toerismevlaanderen\Lodging\Value\PhoneNumbers;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\QualityLabels;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\Registration;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\StarRating;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\WebsiteAddress;
+use DigipolisGent\Toerismevlaanderen\Lodging\Value\WebsiteAddresses;
 use GuzzleHttp\Psr7\Stream;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\ResponseInterface;
@@ -101,19 +104,19 @@ class LodgingHandlerTest extends TestCase
           "type": "typed-literal",
           "datatype": "http://www.w3.org/2001/XMLSchema#float"
         },
-        "contactPoint_phoneNumber": {
+        "contactPoint_phoneNumbers": {
           "value": "+32 9 123 12 12",
           "type": "literal"
         },
-        "contactPoint_emailAddress": {
+        "contactPoint_emailAddresses": {
           "value": "info@foo.baz",
           "type": "literal"
         },
-        "contactPoint_websiteAddress": {
+        "contactPoint_websiteAddresses": {
           "value": "https://foo.baz",
           "type": "uri"
         },
-        "starRating": {
+        "rating": {
           "value": "4 *",
           "type": "literal"
         },
@@ -183,9 +186,15 @@ EOT;
                 Registration::fromTypeAndStatus('B&B', 'Foo status'),
                 Address::fromDetails('Foo street', '8', 'b', '9000', 'Foo locality', Coordinates::fromLongitudeLatitude(1.234, 56.789)),
                 ContactInfo::fromDetails(
-                    PhoneNumber::fromNumber('+32 9 123 12 12'),
-                    EmailAddress::fromAddress('info@foo.baz'),
-                    WebsiteAddress::fromUrl('https://foo.baz')
+                    PhoneNumbers::fromPhoneNumbers(
+                        PhoneNumber::fromNumber('+32 9 123 12 12')
+                    ),
+                    EmailAddresses::fromEmailAddresses(
+                        EmailAddress::fromAddress('info@foo.baz')
+                    ),
+                    WebsiteAddresses::fromWebsiteAddresses(
+                        WebsiteAddress::fromUrl('https://foo.baz')
+                    )
                 ),
                 StarRating::fromEuropeanFormat('4 *'),
                 QualityLabels::fromLabels('Label 1', 'Label 2'),