+++ tests/Normalizer/FromArray/LodgingNormalizerTest.php
@@ -9,15 +9,19 @@ use DigipolisGent\Toerismevlaanderen\Lodging\Value\Address;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\ContactInfo;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\Coordinates;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\EmailAddress;
+use DigipolisGent\Toerismevlaanderen\Lodging\Value\EmailAddresses;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\Image;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\Images;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\Lodging;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\LodgingId;
+use DigipolisGent\Toerismevlaanderen\Lodging\Value\NoRating;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\PhoneNumber;
+use DigipolisGent\Toerismevlaanderen\Lodging\Value\PhoneNumbers;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\QualityLabels;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\Registration;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\StarRating;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\WebsiteAddress;
+use DigipolisGent\Toerismevlaanderen\Lodging\Value\WebsiteAddresses;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -39,8 +43,7 @@ class LodgingNormalizerTest extends TestCase
             'registration' => [
                 'type' => 'B&B',
                 'status' => 'Erkend',
-            ],
-            'starRating' => '2 *',
+            ]
         ];
 
         $expectedLodging = Lodging::fromDetails(
@@ -51,11 +54,11 @@ class LodgingNormalizerTest extends TestCase
             Registration::fromTypeAndStatus('B&B', 'Erkend'),
             Address::fromDetails('', '', '', '', '', Coordinates::fromLongitudeLatitude(0, 0)),
             ContactInfo::fromDetails(
-                PhoneNumber::withoutNumber(),
-                EmailAddress::withoutAddress(),
-                WebsiteAddress::withoutUrl()
+                PhoneNumbers::fromPhoneNumbers(),
+                EmailAddresses::fromEmailAddresses(),
+                WebsiteAddresses::fromWebsiteAddresses()
             ),
-            StarRating::fromEuropeanFormat('2 *'),
+            NoRating::create(),
             QualityLabels::fromLabels(),
             Images::fromImages()
         );
@@ -93,11 +96,11 @@ class LodgingNormalizerTest extends TestCase
                 ],
             ],
             'contactPoint' => [
-                'phoneNumber' => '+32 9 123 12 12',
-                'emailAddress' => 'foo@biz.baz',
-                'websiteAddress' => 'https://foo.baz',
+                'phoneNumbers' => ['+32 9 123 12 12'],
+                'emailAddresses' => ['foo@biz.baz'],
+                'websiteAddresses' => ['https://foo.baz'],
             ],
-            'starRating' => '2 *',
+            'rating' => '2 *',
             'qualityLabels' => ['Label 1', 'Label 2'],
             'images' => [
                 'http://foo.bar/image/1.jpg',
@@ -113,9 +116,15 @@ class LodgingNormalizerTest extends TestCase
             Registration::fromTypeAndStatus('B&B', 'Erkend'),
             Address::fromDetails('Foo street', '8', 'b', '9000', 'Foo locality', Coordinates::fromLongitudeLatitude(1.234, 56.789)),
             ContactInfo::fromDetails(
-                PhoneNumber::fromNumber('+32 9 123 12 12'),
-                EmailAddress::fromAddress('foo@biz.baz'),
-                WebsiteAddress::fromUrl('https://foo.baz')
+                PhoneNumbers::fromPhoneNumbers(
+                    PhoneNumber::fromNumber('+32 9 123 12 12')
+                ),
+                EmailAddresses::fromEmailAddresses(
+                    EmailAddress::fromAddress('foo@biz.baz')
+                ),
+                WebsiteAddresses::fromWebsiteAddresses(
+                    WebsiteAddress::fromUrl('https://foo.baz')
+                )
             ),
             StarRating::fromEuropeanFormat('2 *'),
             QualityLabels::fromLabels('Label 1', 'Label 2'),