+++ tests/Serializer/ContactInfoArraySerializerTest.php
@@ -7,8 +7,11 @@ namespace DigipolisGent\Tests\Toerismevlaanderen\Normalizer;
 use DigipolisGent\Toerismevlaanderen\Lodging\Serializer\ContactInfoArraySerializer;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\ContactInfo;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\EmailAddress;
+use DigipolisGent\Toerismevlaanderen\Lodging\Value\EmailAddresses;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\PhoneNumber;
+use DigipolisGent\Toerismevlaanderen\Lodging\Value\PhoneNumbers;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\WebsiteAddress;
+use DigipolisGent\Toerismevlaanderen\Lodging\Value\WebsiteAddresses;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -24,15 +27,21 @@ class ContactInfoArraySerializerTest extends TestCase
     public function addressArrayContainsAllData(): void
     {
         $contactInfo = ContactInfo::fromDetails(
-            PhoneNumber::fromNumber('+32 9 123 12 12'),
-            EmailAddress::fromAddress('info@biz.baz'),
-            WebsiteAddress::fromUrl('https://foo.baz')
+            PhoneNumbers::fromPhoneNumbers(
+                PhoneNumber::fromNumber('+32 9 123 12 12')
+            ),
+            EmailAddresses::fromEmailAddresses(
+                EmailAddress::fromAddress('info@biz.baz')
+            ),
+            WebsiteAddresses::fromWebsiteAddresses(
+                WebsiteAddress::fromUrl('https://foo.baz')
+            )
         );
 
         $expectedArray = [
-            'phoneNumber' => '+32 9 123 12 12',
-            'emailAddress' => 'info@biz.baz',
-            'websiteAddress' => 'https://foo.baz',
+            'phoneNumbers' => ['+32 9 123 12 12'],
+            'emailAddresses' => ['info@biz.baz'],
+            'websiteAddresses' => ['https://foo.baz'],
         ];
 
         $serializer = new ContactInfoArraySerializer();