+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 All Notable changes to `digipolisgent/toerismevlaanderen-lodging` package.
 
+## [0.2.0]
+
+### Added
+
+* VG-1660: Added Grumphp to run quality checks.
+
+### Changed
+
+* VG-1660: Replaced the single phone number, email address, website address by
+  collections.
+* VG-1660: Replaced star ratings by rating interface: the returned rating value
+  can also be a category rating or can be empty.
+
+### Fixed
+
+* VG-1660: Fixed the lodging details query.
+
 ## [0.1.0]
 
 Initial package.
@@ -12,5 +29,6 @@ Initial package.
 * Added getting a list of lodgings by given filters.
 * Added getting the details of a single lodging.
 
+[0.2.0]: https://github.com/digipolisgent/php_package_dg-toerismevlaanderen-lodging/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/digipolisgent/php_package_dg-toerismevlaanderen-lodging/releases/tag/0.1.0
 [Unreleased]: https://github.com/digipolisgent/php_package_dg-toerismevlaanderen-lodging/compare/master...develop